Zmieniam test{1,2,3,4} na język C

+++ network_test2.c
@@ -107,7 +107,7 @@ int main()
 
     printf("*****\nALL ASSERTIONS PASSED\n*****\n\n");
 
-    int licznik = 0;
+    unsigned long licznik = 0;
 
     for (c = 'a'; c <= 'z'; c++)
     {
